requirements.txt
PyQt6>=6.6.0
keyboard>=0.13.5
pyautogui>=0.9.54
pystray>=0.19.5
Pillow>=10.1.0
pyperclip>=1.8.2
